requirements.txt
pydantic==2.9.2
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.1
pytest==8.3.4
requests==2.32.3
beautifulsoup4==4.12.3
google-search-results==2.4.2
google-generativeai==0.8.3

